Check if pid file exists before removing it
darcs-hash:20070619191207-86b55-70c379307a268b19e55bffd4203e19351e63c467.gz

+++ src/jcl/runner.py
@@ -194,7 +194,8 @@ class JCLRunner(object):
             run_func()
             self.logger.debug(self.component_name + " is exiting")
         finally:
-            os.remove(self.pid_file)
+            if os.path.exists(self.pid_file):
+                os.remove(self.pid_file)
 
     def run(self):
         def run_func():
